Remove unused rust code
